[XLA:Python] Fix use-after-free in pjit.
The following use-after-free was observed in a free-threaded build of JAX:

https://gist.githubusercontent.com/vfdev-5/d08f1939a560b54b6d34f6c38efdad18/raw/26455d2aa75ff92cff9af9bb2a8e3e38c16b47ac/tests_export_back_compat_test.313.log

I believe what is happening here is that we are attempting to access the function that forms part of the cache key to hash it, but after we've freed the function in question.

Precompute the hash instead, which avoids the problem.

PiperOrigin-RevId: 733820157